feat: refactoring dir databases structure
