+++ src/app.module.ts
@@ -1,8 +1,8 @@
 import { ConfigModule } from "@nestjs/config";
 import { Module } from "@nestjs/common";
-import { DatabaseModule } from "./database";
+import { DatabaseModule } from "./dataAccess/databases";
 import { RecluterModule } from "./routes/index.routes";
-import { MongoDatabaseModule } from "./database/mongo/database.module";
+import { MongoDatabaseModule } from "./dataAccess/databases/mongodb/database.module";
 import { StatsModule } from "./routes/stats/stats.module";
 @Module({
   imports: [
