+++ src/dataAccess/databases/index.ts
@@ -1,0 +1,2 @@
+export * from "./postgresql/database.module";
+export * from "./mongodb/database.module"
