+++ src/dataAccess/databases/postgresql/entities/index.ts
@@ -1,0 +1,1 @@
+export * from './sequenceDNA.entity'
