+++ src/dataAccess/databases/postgresql/entities/sequenceDNA.entity.ts
