+++ src/dataAccess/databases/postgresql/index.ts
@@ -1,0 +1,3 @@
+export * from './entities'
+export * from './timestamp'
+export * from './repositories'
