--- src/database/postgreSQL/index.ts
@@ -1,2 +1,0 @@
-export * from './entities/sequenceDNA.entity'
-export * from './timestamp'
