+++ src/routes/stats/stats.module.ts
@@ -3,12 +3,12 @@ import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { StatsController } from './stats.controller';
 import { StatsService } from './stats.service';
-import { SequenceDNARepository } from 'src/database/repositoriesPostgreSQL';
+import { SequenceDNARepository } from 'src/dataAccess/databases/postgresql/repositories';
 import { ExceptionsModule } from 'src/configurations/exceptions';
 import { MongooseModule } from '@nestjs/mongoose';
-import { DNASequenceMongo, DNASequenceSchema } from 'src/database/mongo';
-import { DNASequenceMongoRepository, StatsMongoReqpository } from 'src/database/repositoriesMongoDB';
-import { StatsSequencesMongo, StatsSequencesSchema } from 'src/database/mongo/schemas/Stats.schema';
+import { DNASequenceMongo, DNASequenceSchema } from 'src/dataAccess/databases/mongodb';
+import { DNASequenceMongoRepository, StatsMongoReqpository } from 'src/dataAccess/databases/mongodb/repositories';
+import { StatsSequencesMongo, StatsSequencesSchema } from 'src/dataAccess/databases/mongodb/schemas/Stats.schema';
 
 @Module({
   imports: [
