+++ src/tests/recruter/recruter.service.spec.ts
@@ -2,8 +2,8 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { RecluterService } from 'src/routes/recluter/recluter.service';
 import { getModelToken } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { DNASequenceMongo  } from 'src/database/mongo/schemas/DNA.schema';
-import { DNASequenceMongoRepository } from 'src/database/repositoriesMongoDB';
+import { DNASequenceMongo  } from 'src/dataAccess/databases/mongodb/schemas/DNA.schema';
+import { DNASequenceMongoRepository } from 'src/dataAccess/databases/mongodb/repositories';
 
 describe('recruterService', () => {
   let recruterService: RecluterService;
