+++ src/tests/stats/stats.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from "@nestjs/testing";
-import { StatsSequencesMongo } from "src/database/mongo/schemas/Stats.schema";
-import { StatsMongoReqpository } from "src/database/repositoriesMongoDB";
+import { StatsSequencesMongo } from "src/dataAccess/databases/mongodb/schemas/Stats.schema";
+import { StatsMongoReqpository } from "src/dataAccess/databases/mongodb/repositories";
 import { StatsController } from "src/routes/stats/stats.controller";
 import { StatsService } from "src/routes/stats/stats.service";
